Application/imitagram/requirements.txt
Django==2.1.1
django-allauth==0.37.1
django-rest-auth==0.9.3
djangorestframework==3.8.2
Pillow==5.2.0
